landing(fix): move all components to directory
