+++ src/pages/landing.tsx
@@ -1,6 +1,6 @@
 import ROUTES from '../routes'
 import styled from 'styled-components'
-import FaqPage from '../components/Faq'
+import FaqPage from '../components/landing/Faq'
 import Header from '../components/headers/Header'
 import Footer from '../components/Footer'
 import { textSerif } from '../style/utils'
@@ -8,12 +8,12 @@ import { CIRCLE_SIZE } from '../constants'
 import { useAuthStore } from '../store/auth'
 import { useNavigate } from 'react-router-dom'
 import { TextSection } from '../components/Layout'
-import Explanation from '../components/Explanation'
+import Explanation from '../components/landing/Explanation'
 import { PrimaryButton } from '../components/Button'
 import { Trans, useTranslation } from 'react-i18next'
 import LandingBg from '../assets/landing-boarder.png'
 import { useCallback, useRef, useEffect } from 'react'
-import LatestRecords from '../components/LatestRecords'
+import LatestRecords from '../components/landing/LatestRecords'
 import { Description, PageTitle } from '../components/Text'
 import { isMobile } from '../utils'
